# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "YU-GI-OH! Api Rest",
        "description": "Implementation of Yu-Gi-Oh! Api Rest",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local enviroment"
        }
    ],
    "paths": {
...
        "/api/images": {
            "get": {
                "tags": [
                    "Images"
                ],
                "summary": "List all images.",
                "operationId": "image-index",
                "responses": {
                    "200": {
                        "description": "Get an image collection",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ImageResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Images"
                ],
                "summary": "Create an image.",
                "operationId": "image-store",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "name",
                                    "extension",
                                    "file"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "File name",
                                        "type": "string"
                                    },
                                    "extension": {
                                        "description": "File extension",
                                        "type": "string"
                                    },
                                    "file": {
                                        "description": "Base64 code",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Image created.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImageResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                }
            }
        },
        "/api/image/{id}": {
            "get": {
                "tags": [
                    "Images"
                ],
                "summary": "Get a specific image.",
                "operationId": "image-show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Image ID.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Image obteined.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImageResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Images"
                ],
                "summary": "Delete an image.",
                "operationId": "image-delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Image ID.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Image deleted."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                }
            }
        },
        "/api/images/{id}": {
            "patch": {
                "tags": [
                    "Images"
                ],
                "summary": "Create an image.",
                "operationId": "image-update",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "File name",
                                        "type": "string"
                                    },
                                    "extension": {
                                        "description": "File extension",
                                        "type": "string"
                                    },
                                    "file": {
                                        "description": "Base64 code",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Image updated.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ImageResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request."
                    },
                    "404": {
                        "description": "Not found."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ImageResource": {
                "title": "Image",
                "description": "Card's image",
                "properties": {
                    "id": {
                        "description": "Image's ID",
                        "type": "integer",
                        "format": ""
                    },
                    "name": {
                        "description": "Fila name",
                        "type": "string",
                        "format": ""
                    },
                    "extension": {
                        "description": "File's extension",
                        "type": "string",
                        "format": ""
                    },
                    "file": {
                        "description": "Base 64 code",
                        "type": "string",
                        "format": ""
                    }
                },
                "type": "object",
                "example": {
                    "id": 1,
                    "name": "Dark Margician",
                    "extension": "jpg",
                    "file": "TWFuIGlzIGRpc3Rpbmd1aXNoZWQTWFuIGlzIGRpc3Rpbmd1aXNoZWQTWFuIGlzIGRpc3Rpbmd1aXNoZWQTWFuIGlzIGRpc3Rpbmd1aXNoZWQ"
                }
            },
...
        }
    }
}